Try to import MicroPython u-modules. Drop support of CPython < 2.6.
Most built-in modules in MicroPython can be imported with a preceding 'u' to express their reduced abilities, and (in the Unix port) to distinguish them from their fully implemented variants.
Importing from __future__ must be done at the very beginning (even before 'try'), but raises ImportError in MicroPython and therefore must be removed.