Quick fix for "with assertRaises()" in MicroPython.
The statement "with assertRaises(errtype) as ctxt" checks the type of a raised exception, but in MicroPython it does not store the exception into ctxt like in CPython.

Check here if it has been fixed upstream: https://github.com/micropython/micropython-lib/pull/280
If yes then this fix can be removed, but does not hurt if it stays.